+++ bugsnag-android-core/src/main/java/com/bugsnag/android/ErrorStore.java
@@ -145,20 +145,27 @@ class ErrorStore extends FileStore<Error> {
 
     private void flushErrorReport(File errorFile) {
         try {
-            Error error = ErrorReader.readError(config, errorFile);
-            Report report = new Report(config.getApiKey(), error);
+            Report report;
 
-            for (BeforeSend beforeSend : config.getBeforeSendTasks()) {
-                try {
-                    if (!beforeSend.run(report)) {
-                        deleteStoredFiles(Collections.singleton(errorFile));
-                        Logger.info("Deleting cancelled error file " + errorFile.getName());
-                        return;
+            if (config.getBeforeSendTasks().isEmpty()) {
+                report = new Report(config.getApiKey(), errorFile);
+            } else {
+                Error error = ErrorReader.readError(config, errorFile);
+                report = new Report(config.getApiKey(), error);
+
+                for (BeforeSend beforeSend : config.getBeforeSendTasks()) {
+                    try {
+                        if (!beforeSend.run(report)) {
+                            deleteStoredFiles(Collections.singleton(errorFile));
+                            Logger.info("Deleting cancelled error file " + errorFile.getName());
+                            return;
+                        }
+                    } catch (Throwable ex) {
+                        Logger.warn("BeforeSend threw an Exception", ex);
                     }
-                } catch (Throwable ex) {
-                    Logger.warn("BeforeSend threw an Exception", ex);
                 }
             }
+
             config.getDelivery().deliver(report, config);
 
             deleteStoredFiles(Collections.singleton(errorFile));